requirements.txt
asgiref==3.8.1
astroid==3.3.9
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coverage==7.8.0
dill==0.4.0
Django==5.2
django-extensions==4.1
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.16.0
djangorestframework-stubs==3.15.3
flake8==7.2.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.10.1
pylint==3.3.6
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
requests==2.32.3
sqlparse==0.5.3
tomlkit==0.13.2
types-PyYAML==6.0.12.20250402
types-requests==2.32.0.20250328
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
